requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
python-dotenv==0.21.0
psycopg2
requests~=2.28.1
loguru
pillow
